Remove class structure from template validation
